+++ partiql-logical-planner/src/lower.rs
@@ -34,7 +34,7 @@ use partiql_ast_passes::error::{AstTransformError, AstTransformationError};
 use partiql_catalog::Catalog;
 use partiql_extension_ion::decode::{IonDecoderBuilder, IonDecoderConfig};
 use partiql_extension_ion::Encoding;
-use partiql_logical::AggFunc::{AggAvg, AggCount, AggMax, AggMin, AggSum};
+use partiql_logical::AggFunc::{AggAny, AggAvg, AggCount, AggEvery, AggMax, AggMin, AggSum};
 use std::sync::atomic::{AtomicU32, Ordering};
 
 type FnvIndexMap<K, V> = IndexMap<K, V, FnvBuildHasher>;
@@ -747,8 +747,13 @@ impl<'a, 'ast> Visitor<'ast> for AstToLogical<'a> {
         Traverse::Continue
     }
 
-    fn enter_select(&mut self, _select: &'ast Select) -> Traverse {
-        Traverse::Continue
+    fn enter_select(&mut self, select: &'ast Select) -> Traverse {
+        if select.having.is_some() && select.group_by.is_none() {
+            self.errors.push(AstTransformError::HavingWithoutGroupBy);
+            Traverse::Stop
+        } else {
+            Traverse::Continue
+        }
     }
 
     fn exit_select(&mut self, _select: &'ast Select) -> Traverse {
@@ -1024,7 +1029,7 @@ impl<'a, 'ast> Visitor<'ast> for AstToLogical<'a> {
         let call_expr = self
             .fnsym_tab
             .lookup(&name)
-            .map(|cd| call_def_to_vexpr(cd))
+            .map(call_def_to_vexpr)
             .or_else(|| {
                 self.catalog
                     .get_function(&name)
@@ -1210,6 +1215,18 @@ impl<'a, 'ast> Visitor<'ast> for AstToLogical<'a> {
                 name: new_name,
                 expr: arg,
                 func: AggSum,
+                setq,
+            },
+            "any" | "some" => AggregateExpression {
+                name: new_name,
+                expr: arg,
+                func: AggAny,
+                setq,
+            },
+            "every" => AggregateExpression {
+                name: new_name,
+                expr: arg,
+                func: AggEvery,
                 setq,
             },
             _ => {